+++ stag_lib/CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to the library are documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.1] - 2023-03-02
 ### Fixed
 - Reduced the memory usage of the SBM methods
+- Correct the SBM approximate sampling distribution
 
 ### Added
 - Added signless Laplacian methods to the graph object